requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
docutils==0.21.2
fake-useragent==1.5.1
h11==0.14.0
idna==3.7
lockfile==0.12.2
lxml==5.2.1
outcome==1.3.0.post0
pid==3.0.4
PySocks==1.7.1
python-daemon==3.0.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.20.0
service==0.6.0
setproctitle==1.3.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
